Move hash_from_dict and make it better support OrderedDict
