+++ theano/gof/tests/test_utils.py
@@ -1,5 +1,7 @@
 import theano
-from theano.gof.utils import give_variables_names, unique, remove
+from theano.compat import OrderedDict
+from theano.gof.utils import (
+    give_variables_names, hash_from_dict, remove, unique)
 
 
 def test_give_variables_names():
@@ -44,3 +46,20 @@ def test_remove():
     # The list are needed as with python 3, remove and filter return generators
     # and we can't compare generators.
     assert list(remove(even, range(5))) == list(filter(odd, range(5)))
+
+
+def test_hash_from_dict():
+    dicts = [{}, {0: 0}, {0: 1}, {1: 0}, {1: 1},
+             {0: (0,)}, {0: [1]},
+             {0: (0, 1)}, {0: [1, 0]},
+         ]
+    for elem in dicts[:]:
+        dicts.append(OrderedDict(elem))
+    hashs = []
+    for idx, d in enumerate(dicts):
+        h = hash_from_dict(d)
+        assert h not in hashs
+        hashs.append(h)
+
+    # List are not hashable. So they are transformed into tuple.
+    assert hash_from_dict({0: (0,)}) == hash_from_dict({0: [0]})
